♻️ refactor: 이벤트 관련 컨트롤러 리팩토링 및 분리 (#160)
